Change the default sticky note to have transparent border
In this video: https://youtu.be/fXGcOWycgG4?t=1923
it is mentioned that the default border color for the "Convert selected
text elements to sticky notes" script is a transparent border, however
that is not the current default value.

I think setting `transparent` as the default is more useful than
`#000000` because `transparent` can only be set via a script (at least
without affecting the text color) while `#000000` can be set since it is
also a common color for text.